`border` and `borderColor` functions
`border` creates the border for given page. It also supports creating
the border with rounded corner and margin.

Examples:

```
border(5);
border(5, margin: 10, radius: 7);
```

Also supports adjusting a specific margin or radius: `marginTop`,
`marginBottom`, `marginLeft`, `marginRight`, `radiusTopLeft`,
`radiusTopRight`, `radiusBottomLeft` and `radiusBottomRight`.

Border color can be customized using `borderColor` or `color` argument.

```
border(5, color: RGBA.parse("blue"));

// Same as above
borderColor("blue");
border(5);
```

+++ source/chitra/package.d
@@ -7,11 +7,11 @@ public
 {
     import chitra.helpers;
     import chitra.elements.formatted_strings;
+    import chitra.rgba;
 }
 import chitra.context;
 import chitra.properties;
 import chitra.elements;
-import chitra.rgba;
 
 class Chitra : Context
 {
@@ -47,7 +47,7 @@ class Chitra : Context
        auto ctx = new Chitra(1600, 900);
        ---
      */
-    this(int width = defaultWidth, int height = 0)
+    this(double width = defaultWidth, double height = 0)
     {
         super(width, height);
     }
